avocado: Replace --multiplex with yaml_to_mux plugin
This complex patch replaces the custom `--multiplex` handling with a
brand new plugin called `yaml_to_mux`. This touches several place and
removes the `yaml` functionality from the avocado core and moves it into
the plugin code.

There are still 2 references from plugins to the `yaml_to_mux` plugin
which are not defined as a public API yet.

1. remote runner - which allows copying --multiplex files to remote
                   machine.
2. replay - which allows overriding of the job's mux variants with
            the --multiplex files. Note that by doing this you
            lose the system-wide options. It is described in the
            warning message.
