+++ avocado/plugins/multiplex.py
@@ -16,7 +16,6 @@ import logging
 import sys
 
 from avocado.core import exit_codes, output
-from avocado.core import multiplexer
 from avocado.core import tree
 from avocado.core.plugin_interfaces import CLICmd
 from avocado.core.settings import settings
@@ -36,9 +35,6 @@ class Multiplex(CLICmd):
 
     def configure(self, parser):
         parser = super(Multiplex, self).configure(parser)
-        if multiplexer.MULTIPLEX_CAPABLE:
-            parser.add_argument("-m", '--multiplex', nargs='*',
-                                help='Path(s) to a multiplex file(s)')
 
         parser.add_argument('--filter-only', nargs='*', default=[],
                             help='Filter only path(s) from multiplexing')
@@ -76,8 +72,7 @@ class Multiplex(CLICmd):
         if err:
             log.error(err)
             sys.exit(exit_codes.AVOCADO_FAIL)
-        mux = multiplexer.Mux(args.mux_debug)
-        root = mux.data
+        mux = args.mux
         try:
             mux.parse(args)
         except (IOError, ValueError) as details:
@@ -92,7 +87,7 @@ class Multiplex(CLICmd):
                 verbose += 2
             use_utf8 = settings.get_value("runner.output", "utf8",
                                           key_type=bool, default=None)
-            log.debug(tree.tree_view(root, verbose, use_utf8))
+            log.debug(tree.tree_view(mux.variants.root, verbose, use_utf8))
             sys.exit(exit_codes.AVOCADO_ALL_OK)
 
         log.info('Variants generated:')
