+++ docs/source/Mux.rst
@@ -1,13 +1,100 @@
-.. _multiplex_configuration:
+.. _mux:
 
-=======================
-Multiplex Configuration
-=======================
+===================
+Test variants - Mux
+===================
+
+The ``Mux`` is a special mechanism to produce multiple variants of the same
+test with different parameters. This is essential in order to get a decent
+coverage and avocado allows several ways to define those parameters from
+simple enumeration of key/value pairs to complex trees which allows in simple
+manner define test matrices with all possible variants.
+
+This sounds similar to sparse matrix jobs in Jenkins, but the difference is
+that instead of filters, which are available too, avocado allows specifying
+so called ``mux domains``, which is a nicer way to represent data.
+As the data is represented in trees it creates all possible variants
+per domain and then all combinations of these. It sounds complicated, but
+in reality it follows the way people are used to define dependencies,
+therefor it's very simple to use and clear even in complex cases.
+
+The best explanation comes usually from examples, so feel free to scroll down
+to `yaml_to_mux plugin`_ section, which uses the default mux plugin to feed
+the Mux.
+
+
+Mux internals
+-------------
+
+The ``Mux`` is a core part of avocado and one can see it as a ``multiplexed``
+database, which contains key/value pairs associated to given paths and
+as we are talking about a tree of those, we call the paths ``Nodes``.
+
+Mux allows iterating through all possible combinations which are stored in
+the database, which is called ``multiplexation``. Mux yields ``variants``,
+which are lists of leaf nodes with their values, which are then processed
+into ``AvocadoParams``. Those params are available in tests as
+``self.params`` and one can query for the current parameters::
+
+    self.params.get(key="my_key", path="/some/location/*",
+                    default="default_value")
+
+Let's get back to Mux for a while. As mentioned earlier, it's a database
+which allows storing multiple variants of test parameters. To fill the
+database, you can use several commands.
+
+1. ``--mux-inject`` - injects directly [path:]key:node values from the
+   cmdline (see ``avocado multiplex -h``)
+2. ``yaml_to_mux plugin`` - allows parsing ``yaml`` files into the Mux
+   database (see `yaml_to_mux plugin`_)
+3. Custom plugin using the simple ``Mux`` API (see `mux_api`_)
+
+
+.. _mux_api:
+
+Mux API
+-------
+
+.. warning:: This API is internal, we might change it at any moment. On the
+             other hand we maintain ``avocado-virt`` plugin which uses this
+             API so in such case we'd provide a patch there demonstrating
+             the necessary changes.
+
+The ``Mux`` object is defined in ``avocado/core/multiplexer.py``, is always
+instantiated in ``avocado.core.parser.py`` and always available in
+``args.mux``. The basic workflow is:
+
+1. Initialize ``Mux`` in ``args.mux``
+2. Fill it with data (``plugins`` or ``job``)
+3. Multiplex it (in ``job``)
+4. Iterate through all variants on all job's tests
+
+Once the ``Mux`` object is multiplexed (3), it's restricted to alter the
+data (2) to avoid changing the already produced data.
+
+The main API needed for your plugins, which we are going to try keeping as
+stable as possible is:
+
+* mux.is_parsed() - to find out whether the object was already parsed
+* data_inject(key, value, path=None) - to inject key/value pairs optionaly
+  to a given path (by default '/')
+* data_merge(tree) - to merge ``avocado.core.tree.TreeNode``-like tree
+  into the database.
+
+Given these you should be able to implement any kind of parser or params
+feeder, should you require one. We favor ``yaml`` and therefor we implemented
+a ``yaml_to_mux`` plugin which can be found in
+``avocado/plugins/yaml_to_mux.py`` and on it we also describe the way
+``Mux`` works: `yaml_to_mux plugin`_
+
+
+Yaml_to_mux plugin
+==================
 
 In order to get a good coverage one always needs to execute the same test
 with different parameters or in various environments. Avocado uses the
-term ``Multiplexation`` to generate multiple variants of the same test with
-different values. To define these variants and values
+term ``Multiplexation`` or ``Mux`` to generate multiple variants of the same
+test with different values. To define these variants and values
 `YAML <http://www.yaml.org/>`_ files are used. The benefit of using YAML
 file is the visible separation of different scopes. Even very advanced setups
 are still human readable, unlike traditional sparse, multi-dimensional-matrices
@@ -48,7 +135,7 @@ flags (lines 2, 9, 14, 19) which modifies the behavior.
 
 
 Nodes
-=====
+-----
 
 They define context of the key=>value pairs allowing us to easily identify
 for what this values might be used for and also it makes possible to define
@@ -85,7 +172,7 @@ parameters available in tests.
 
 
 Keys and Values
-===============
+---------------
 
 Every value other than dict (4,6,8,11) is used as value of the antecedent
 node.
@@ -127,7 +214,7 @@ This means that the value can be of any YAML supported value, eg. bool, None,
 list or custom type, while the key is always string.
 
 Variants
-========
+--------
 
 In the end all leaves are gathered and turned into parameters, more specifically into
 ``AvocadoParams``::
@@ -182,7 +269,7 @@ Results in::
 
 
 Resolution order
-================
+----------------
 
 You can see that only leaves are part of the test parameters. It might happen
 that some of these leaves contain different values of the same key. Then
@@ -225,7 +312,7 @@ relative paths (the ones starting with ``*``)
 
 
 Injecting files
-===============
+---------------
 
 You can run any test with any YAML file by::
 
@@ -272,7 +359,7 @@ parameters.
 
 
 Multiple files
-==============
+--------------
 
 You can provide multiple files. In such scenario final tree is a combination
 of the provided files where later nodes with the same name override values of
@@ -318,7 +405,7 @@ Whole file is **merged** into the node where it's defined.
 
 
 Advanced YAML tags
-==================
+------------------
 
 There are additional features related to YAML files. Most of them require values
 separated by ``":"``. Again, in all such cases it's mandatory to add a white space
@@ -391,7 +478,7 @@ child, etc. Example is in section `Variants`_
 
 
 Complete example
-================
+----------------
 
 Let's take a second look at the first example::
