+++ selftests/unit/test_multiplexer.py
@@ -4,6 +4,7 @@ import sys
 
 from avocado.core import multiplexer
 from avocado.core import tree
+from avocado.plugins import yaml_to_mux
 
 if sys.version_info[:2] == (2, 6):
     import unittest2 as unittest
@@ -24,54 +25,48 @@ def combine(leaves_pools):
 
 class TestMultiplex(unittest.TestCase):
 
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE,
+                     "Not multiplex capable")
     def setUp(self):
-        self.mux_tree = tree.create_from_yaml(['/:' + PATH_PREFIX +
-                                               'examples/mux-selftest.yaml'])
+        self.mux_tree = yaml_to_mux.create_from_yaml(['/:' + PATH_PREFIX +
+                                                      'examples/mux-selftest.'
+                                                      'yaml'])
         self.mux_full = tuple(multiplexer.MuxTree(self.mux_tree))
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_empty(self):
         act = tuple(multiplexer.MuxTree(tree.TreeNode()))
         self.assertEqual(act, (['', ],))
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_partial(self):
         exp = (['intel', 'scsi'], ['intel', 'virtio'], ['amd', 'scsi'],
                ['amd', 'virtio'], ['arm', 'scsi'], ['arm', 'virtio'])
         act = tuple(multiplexer.MuxTree(self.mux_tree.children[0]))
         self.assertEqual(act, exp)
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_full(self):
         self.assertEqual(len(self.mux_full), 12)
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_create_variants(self):
-        from_file = multiplexer.yaml2tree(
+        from_file = yaml_to_mux.create_from_yaml(
             ["/:" + PATH_PREFIX + 'examples/mux-selftest.yaml'])
         from_file = multiplexer.MuxTree(from_file)
         self.assertEqual(self.mux_full, tuple(from_file))
 
     # Filters are tested in tree_unittests, only verify `multiplex_yamls` calls
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_filter_only(self):
         exp = (['intel', 'scsi'], ['intel', 'virtio'])
-        act = multiplexer.yaml2tree(["/:" + PATH_PREFIX +
-                                     'examples/mux-selftest.yaml'],
-                                    ('/hw/cpu/intel',
-                                     '/distro/fedora',
-                                     '/hw'))
+        act = yaml_to_mux.create_from_yaml(["/:" + PATH_PREFIX +
+                                            'examples/mux-selftest.yaml'])
+        act = tree.apply_filters(act, ('/hw/cpu/intel', '/distro/fedora',
+                                       '/hw'))
         act = tuple(multiplexer.MuxTree(act))
         self.assertEqual(act, exp)
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_filter_out(self):
-        act = multiplexer.yaml2tree(["/:" + PATH_PREFIX +
-                                     'examples/mux-selftest.yaml'],
-                                    None,
-                                    ('/hw/cpu/intel',
-                                     '/distro/fedora',
-                                     '/distro'))
+        act = yaml_to_mux.create_from_yaml(["/:" + PATH_PREFIX +
+                                            'examples/mux-selftest.yaml'])
+        act = tree.apply_filters(act, None, ('/hw/cpu/intel', '/distro/fedora',
+                                             '/distro'))
         act = tuple(multiplexer.MuxTree(act))
         self.assertEqual(len(act), 4)
         self.assertEqual(len(act[0]), 3)
@@ -85,8 +80,8 @@ class TestMultiplex(unittest.TestCase):
 class TestAvocadoParams(unittest.TestCase):
 
     def setUp(self):
-        yamls = multiplexer.yaml2tree(["/:" + PATH_PREFIX +
-                                       'examples/mux-selftest-params.yaml'])
+        yamls = yaml_to_mux.create_from_yaml(["/:" + PATH_PREFIX +
+                                              'examples/mux-selftest-params.yaml'])
         self.yamls = iter(multiplexer.MuxTree(yamls))
         self.params1 = multiplexer.AvocadoParams(self.yamls.next(), 'Unittest1',
                                                  ['/ch0/*', '/ch1/*'], {})
@@ -95,13 +90,13 @@ class TestAvocadoParams(unittest.TestCase):
         self.params2 = multiplexer.AvocadoParams(self.yamls.next(), 'Unittest2',
                                                  ['/ch1/*', '/ch0/*'], {})
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_pickle(self):
         params = pickle.dumps(self.params1, 2)  # protocol == 2
         params = pickle.loads(params)
         self.assertEqual(self.params1, params)
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_basic(self):
         self.assertEqual(self.params1, self.params1)
         self.assertNotEqual(self.params1, self.params2)
@@ -110,7 +105,7 @@ class TestAvocadoParams(unittest.TestCase):
         str(multiplexer.AvocadoParams([], 'Unittest', [], {}))
         self.assertEqual(15, sum([1 for _ in self.params1.iteritems()]))
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_unhashable(self):
         """ Verifies that unhashable arguments can be passed to params.get """
         self.assertEqual(self.params1.get("root", "/ch0/", ["foo"]), ["foo"])
@@ -118,7 +113,7 @@ class TestAvocadoParams(unittest.TestCase):
                                           '/ch0/ch0.1/ch0.1.1/ch0.1.1.1/',
                                           ['bar']), 'unique1')
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_get_abs_path(self):
         # /ch0/ is not leaf thus it's not queryable
         self.assertEqual(self.params1.get('root', '/ch0/', 'bbb'), 'bbb')
@@ -140,7 +135,7 @@ class TestAvocadoParams(unittest.TestCase):
                                           '/ch0/ch0.1/ch0.1.1/ch0.1.1.1/',
                                           'hhh'), 'hhh')
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_get_greedy_path(self):
         self.assertEqual(self.params1.get('unique1', '/*/*/*/ch0.1.1.1/',
                                           111), 'unique1')
@@ -164,7 +159,7 @@ class TestAvocadoParams(unittest.TestCase):
         # path matches nothing
         self.assertEqual(self.params1.get('root', '', 999), 999)
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_get_rel_path(self):
         self.assertEqual(self.params1.get('root', default='iii'), 'root')
         self.assertEqual(self.params1.get('unique1', '*', 'jjj'), 'unique1')
@@ -179,7 +174,7 @@ class TestAvocadoParams(unittest.TestCase):
         self.assertEqual(self.params2.get('unique1', '*/ch0.1.1.1/', 'ooo'),
                          'ooo')
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_get_clashes(self):
         # One inherited, the other is new
         self.assertRaisesRegexp(ValueError, r"'clash1'.* \['/ch0/ch0.1/ch0.1.1"
