+++ selftests/unit/test_tree.py
@@ -7,6 +7,7 @@ else:
     import unittest
 
 from avocado.core import tree
+from avocado.plugins import yaml_to_mux
 
 if __name__ == "__main__":
     PATH_PREFIX = "../../../../"
@@ -16,8 +17,8 @@ else:
 
 class TestTree(unittest.TestCase):
     # Share tree with all tests
-    tree = tree.create_from_yaml(['/:' + PATH_PREFIX +
-                                  'examples/mux-selftest.yaml'])
+    tree = yaml_to_mux.create_from_yaml(['/:' + PATH_PREFIX +
+                                         'examples/mux-selftest.yaml'])
 
     def test_node_order(self):
         self.assertIsInstance(self.tree, tree.TreeNode)
@@ -160,8 +161,9 @@ class TestTree(unittest.TestCase):
                          tree2.children[0].children[2].children[1].value)
 
     def test_advanced_yaml(self):
-        tree2 = tree.create_from_yaml(['/:' + PATH_PREFIX + 'examples/mux-'
-                                       'selftest-advanced.yaml'])
+        tree2 = yaml_to_mux.create_from_yaml(['/:' + PATH_PREFIX +
+                                              'examples/mux-selftest-advanced.'
+                                              'yaml'])
         exp = ['intel', 'amd', 'arm', 'scsi', 'virtio', 'fedora', '6',
                '7', 'gentoo', 'mint', 'prod', 'new_node', 'on']
         act = tree2.get_leaves()
